Move RowJsonException out of RowJsonSerializer
